requirements/base.txt
Django==5.0.4
bpython==0.24
channels==4.0.0
channels_redis==4.2.0
daphne==4.1.2
django-cacheops==7.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.1
pytz==2024.1
websocket-client==1.8.0
websockets==11.0
